userService unit test

+++ src/user/user.service.ts
@@ -1,27 +1,61 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
+import * as bcrypt from 'bcrypt';
+import { envVariableKeys } from 'src/common/const/env.const';
+import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    private readonly configService: ConfigService,
   ) {}
 
-  create(createUserDto: CreateUserDto) {
-    return this.userRepository.save(createUserDto);
+  async create(createUserDto: CreateUserDto) {
+    const { email, password } = createUserDto;
+
+    const user = await this.userRepository.findOne({
+      where: {
+        email,
+      },
+    });
+
+    if (user) {
+      throw new BadRequestException('이미 가입한 유저입니다.');
+    }
+
+    const hash = await bcrypt.hash(
+      password,
+      this.configService.get<number>(envVariableKeys.hasRounds),
+    );
+
+    await this.userRepository.save({
+      email,
+      password: hash,
+    });
+
+    return this.userRepository.findOne({
+      where: {
+        email,
+      },
+    });
   }
 
   findAll() {
     return this.userRepository.find();
   }
 
-  findOne(id: number) {
-    const user = this.userRepository.findOne({ where: { id } });
+  async findOne(id: number) {
+    const user = await this.userRepository.findOne({ where: { id } });
 
     if (!user) {
       throw new NotFoundException('존재하지 않은 사용자 입니다.');
@@ -31,18 +65,35 @@ export class UserService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
-    const user = this.userRepository.findOne({ where: { id } });
+    const { password } = updateUserDto;
+    const user = await this.userRepository.findOne({ where: { id } });
 
     if (!user) {
       throw new NotFoundException('존재하지 않은 사용자 입니다.');
     }
 
-    await this.userRepository.update(id, updateUserDto);
+    const hash = await bcrypt.hash(
+      password,
+      this.configService.get<number>(envVariableKeys.hasRounds),
+    );
+
+    await this.userRepository.update(id, {
+      ...updateUserDto,
+      password: hash,
+    });
 
     return this.userRepository.findOne({ where: { id } });
   }
 
-  remove(id: number) {
-    return this.userRepository.delete(id);
+  async remove(id: number) {
+    const user = await this.userRepository.findOne({ where: { id } });
+
+    if (!user) {
+      throw new NotFoundException('존재하지 않은 사용자 입니다.');
+    }
+
+    await this.userRepository.delete(id);
+
+    return id;
   }
 }
